Fixing aimsun.py

+++ flow/core/kernel/network/aimsun.py
@@ -262,7 +262,8 @@ class AimsunKernelNetwork(BaseKernelNetwork):
         cur_dir = os.path.join(config.PROJECT_PATH,
                                'flow/core/kernel/network')
         os.remove(os.path.join(cur_dir, 'data_%s.json' % self.sim_params.port))
-        os.remove('%s_%s' % (self.network.net_params.template, self.sim_params.port))
+        if self.network.net_params.template:
+            os.remove('%s_%s' % (self.network.net_params.template, self.sim_params.port))
 
     ###########################################################################
     #                        State acquisition methods                        #
